csapp: updates to compile and run chapter 5 exercises

+++ csapp/ch05/5.18.c
@@ -59,7 +59,7 @@ int main(void)
     double a[n];
     memset(a, 0, sizeof(double) * n);
 
-    for (size_t i = 0; i < n; i++) {
+    for (int i = 0; i < n; i++) {
         a[i] = i + 1;
     }
